[libs] E: Moving Safe Wrappers for File
